# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * EduCRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}