# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * EduCRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  source?: string;
  notes?: string;
}