# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * EduCRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DealsByStage } from './dealsByStage';

export interface DashboardSummary {
  totalContacts: number;
  openDeals: number;
  totalPipelineValue: number;
  dealsWonThisMonth: number;
  tasksDueToday: number;
  tasksOverdue: number;
  dealsByStage: DealsByStage[];
}