# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * EduCRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DealStatus } from './dealStatus';

export interface Deal {
  id: number;
  title: string;
  value: number;
  status: DealStatus;
  stageId: number;
  contactId: number;
  /** @nullable */
  contactName?: string | null;
  createdAt: Date;
  updatedAt: Date;
}