# lib/api-zod/src/generated/types/dealInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * EduCRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DealStatus } from './dealStatus';

export interface DealInput {
  /** @minLength 1 */
  title: string;
  value: number;
  stageId: number;
  contactId: number;
  status?: DealStatus;
}