# lib/api-zod/src/generated/types/dealList.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * EduCRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Deal } from './deal';

export type DealList = Deal[];